I imported Balkon's original mod
Now I am working on the errors, and figuring out what I need to do in
order to get this working with forge.  Also I need to figure out how to
make sure eclipse is considering PlayerAPI which is a mod that this one
uses.  I added the PlayerAPI to my referenced libraries, and when I run
my package, it says PlayerAPI is loaded in the mods.  But there seem to
be a lot of "player" references which eclipse cannot resolve.  I
honestly have no idea what I am talking about when it comes to
programming, but these are the main things that I have noticed need
fixing.